Fix some windows compile errors in the multiplayer server.

+++ src/multiplayer_server.cpp
@@ -12,7 +12,7 @@
 #include <vector>
 
 #include <iostream>
-#include <inttypes.h>
+#include <cstdint>
 
 #include "asserts.hpp"
 #include "foreach.hpp"
@@ -189,7 +189,7 @@ private:
 				sockets_info_[socket_it->second].udp_endpoint = endpoint;
 
 				GameInfoPtr& game = sockets_info_[socket_it->second].game;
-				if(udp_buf_[0] == 'Z' && game.get() != NULL && !game->started && game->players.size() >= game->nplayers) {
+				if(udp_buf_[0] == 'Z' && game.get() != NULL && !game->started && game->players.size() >= size_t(game->nplayers)) {
 					bool have_sockets = true;
 					foreach(const socket_ptr& sock, game->players) {
 						const SocketInfo& info = sockets_info_[sock];
